Contexts – Error Context: Forward-Compatible Refactor and DomainEvent Integration (#695)
- contexts/error.py: Replace get_error_cmd/GetError with get_error_evt/DomainEvent;
  fix artifact comments (attribute/init); remove LegacyTiferetError union from
  handle_error(); update imports to forward-compatible packages
- contexts/tests/test_error.py: Rename fixture to get_error_evt_mock (DomainEvent);
  update error_context constructor; forward-compatible imports
- assets/constants.py: Rename get_error_cmd → get_error_evt in DEFAULT_SERVICES

Closes #683
