+++ tiferet/assets/constants.py
@@ -39,8 +39,8 @@ DEFAULT_SERVICES = [
         },
     },
     {
-        'service_id': 'get_error_cmd',
-        'module_path': 'tiferet.commands.error',
+        'service_id': 'get_error_evt',
+        'module_path': 'tiferet.events.error',
         'class_name': 'GetError',
     },
     {
